performance optimization for operators detail pages
