--- constants/operators-detail.js
@@ -1,20 +1,0 @@
-const TABS_OPERATORS_DETAIL = [{
-  label: 'overview',
-  value: 'overview'
-}, {
-  label: 'documentation',
-  value: 'documentation',
-  number: '65%'
-}, {
-  label: 'observations',
-  value: 'observations',
-  number: 120
-}, {
-  label: 'operator-detail.tabs.fmus',
-  value: 'fmus',
-  number: 7
-}];
-
-export {
-  TABS_OPERATORS_DETAIL
-};
